requirements.txt
alembic==1.15.2
colorama==0.4.6
colorlog==6.9.0
greenlet==3.2.1
imbalanced-learn==0.12.4
imblearn==0.0
joblib==1.4.2
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.0.2
optuna==4.3.0
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.0
scipy==1.13.1
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
